pyproject.toml
[project]
name = "multi-agent-travel-planning-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cachetools>=7.0.1",
    "chromadb>=1.4.1",
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.61.0",
    "httpx>=0.28.1",
    "langchain>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "lxml>=6.0.2",
    "numpy>=2.4.2",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "timezonefinder>=8.2.1",
    "uvicorn>=0.40.0",
]
